Move duplicate section scan
