+++ src/database/directory.rs
@@ -18,10 +18,11 @@ impl Directory {
     // populates the given issues list with all issues in this directory
     pub fn check(&self, relative_path: &Path, issues: &mut Vec<Issue>) {
         for (path, doc) in &self.docs {
-            //
+            let doc_path = relative_path.join(path);
+            doc.find_duplicate_sections(&doc_path, issues);
         }
         for (path, dir) in &self.dirs {
-            dir.check(relative_path, issues);
+            dir.check(&relative_path.join(path), issues);
         }
     }
